update: use unsigned long

+++ userspace/libsinsp/test/filterchecks/evt.cpp
@@ -76,7 +76,7 @@ TEST_F(sinsp_with_test_input, EVT_FILTER_cmd_str)
 
 	int fd = 1;
 
-	sinsp_evt* evt = add_event_advance_ts(increasing_ts(), 1, PPME_SYSCALL_BPF_2_X, 2, fd, (uint32_t)PPM_BPF_PROG_LOAD);
+	sinsp_evt* evt = add_event_advance_ts(increasing_ts(), 1, PPME_SYSCALL_BPF_2_X, 2, fd, (uint64_t)PPM_BPF_PROG_LOAD);
 
 	ASSERT_EQ(get_field_as_string(evt, "evt.arg.cmd"), "BPF_PROG_LOAD");
 }
